+++ core-ng-api/src/main/java/core/framework/api/validate/Pattern.java
@@ -14,5 +14,7 @@ import static java.lang.annotation.RetentionPolicy.RUNTIME;
 public @interface Pattern {
     String value();
 
+    boolean ignoreCase() default false;
+
     String message() default "field must match /{pattern}/, value={value}";
 }