+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 9.1.3 (9/26/2024 - )
 
+* validation: @Pattern supports @Pattern(ignoreCase = true)
+
 ### 9.1.2 (8/9/2024 - 9/26/2024)
 
 * search: loading from json into search request